# build/swagger.json
{
	"components": {
...
		"schemas": {
...
			"SensorDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"isRunning": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"type",
					"isRunning"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SensorDataDTO": {
				"properties": {
					"sensorName": {
						"type": "string"
					},
					"value": {
						"type": "number",
						"format": "double"
					},
					"timeStamp": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"sensorName",
					"value",
					"timeStamp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SnapshotDownloadResponse": {
				"properties": {
					"newFileName": {
						"type": "string"
					},
					"downloadUrl": {
						"type": "string"
					},
					"startDate": {
						"type": "number",
						"format": "double"
					},
					"endDate": {
						"type": "number",
						"format": "double"
					},
					"decompressionByteLength": {
						"type": "number",
						"format": "double"
					},
					"note": {
						"type": "string"
					}
				},
				"required": [
					"newFileName",
					"downloadUrl",
					"startDate",
					"endDate",
					"decompressionByteLength",
					"note"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LogDTO": {
				"properties": {
					"timeStamp": {
						"type": "number",
						"format": "double"
					},
					"logContent": {
						"type": "string"
					}
				},
				"required": [
					"timeStamp",
					"logContent"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IterableJson": {
				"properties": {},
				"type": "object",
				"additionalProperties": {}
			},
			"Report": {
				"properties": {
					"type": {
						"type": "string",
						"enum": [
							"Ok",
							"Error",
							"Unknown"
						]
					},
					"info": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"warning": {
						"type": "string"
					},
					"values": {
						"$ref": "#/components/schemas/IterableJson"
					}
				},
				"required": [
					"type",
					"info",
					"error",
					"warning"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Timestamp": {
				"properties": {
					"timeStamp": {
						"type": "number",
						"format": "double"
					},
					"isoTimeStamp": {
						"type": "string"
					},
					"normalTimeStamp": {
						"type": "string"
					}
				},
				"required": [
					"timeStamp",
					"isoTimeStamp",
					"normalTimeStamp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CommonEventFormat": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Report"
					},
					{
						"$ref": "#/components/schemas/Timestamp"
					}
				]
			},
			"DatabaseEvent": {
				"properties": {
					"content": {
						"$ref": "#/components/schemas/CommonEventFormat"
					}
				},
				"required": [
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			}
		}
	},
	"info": {
		"title": "lida-api",
		"version": "1.0.0",
		"description": "This is a REST API utilising the CQRS design pattern being exposed via Google Clouds serverless functions.",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/sensor/get": {
			"get": {
				"operationId": "GetSensors",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SensorDTO"
									},
									"type": "array"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"408": {
						"description": "Request Timeout"
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "accessToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/sensor/{typeOrName}/get": {
			"get": {
				"operationId": "GetCategorizedSensors",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SensorDTO"
									},
									"type": "array"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"408": {
						"description": "Request Timeout"
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "accessToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "typeOrName",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/sensor/{name}/data/get": {
			"get": {
				"operationId": "GetSensorData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SensorDataDTO"
									},
									"type": "array"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"408": {
						"description": "Request Timeout"
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "accessToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "startDate",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "endDate",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/sensor/snapshot/{runNumber}/get": {
			"get": {
				"operationId": "GetSensorDataRunSnapshot",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SnapshotDownloadResponse"
									},
									"type": "array"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"408": {
						"description": "Request Timeout"
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "accessToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "runNumber",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/api/v1/log/sensor/get": {
			"get": {
				"operationId": "GetSensorLogs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LogDTO"
									},
									"type": "array"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden"
					},
					"408": {
						"description": "Request Timeout"
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "accessToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/v1/snapshot/sensor/{runNumber}/delete": {
			"delete": {
				"operationId": "DeleteSensorSnapshot",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DatabaseEvent"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"408": {
						"description": "Request Timeout"
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "accessToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "runNumber",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/api/v1/systemCommand/startSystem": {
			"post": {
				"operationId": "StartSystem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DatabaseEvent"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"408": {
						"description": "Request Timeout"
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "accessToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/systemCommand/pauseSystem": {
			"post": {
				"operationId": "PauseSystem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DatabaseEvent"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"408": {
						"description": "Request Timeout"
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "accessToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/systemCommand/stopSystem": {
			"post": {
				"operationId": "StopSystem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DatabaseEvent"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"408": {
						"description": "Request Timeout"
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "accessToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/systemCommand/restartSystem": {
			"post": {
				"operationId": "RestartSystem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DatabaseEvent"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"408": {
						"description": "Request Timeout"
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "accessToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/systemCommand/flags/commit": {
			"post": {
				"operationId": "CommitSystemFlags",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DatabaseEvent"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"408": {
						"description": "Request Timeout"
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "accessToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}